When player is dead cannot pause game
Solves a bug that appears when the game is fading to restart the scene and you pause the game. When the scene is restarted game is freezed without any menu on screen and it won't unfreeze until you click ESC another time.

By telling the game to not pause when the player is dead solves this issue.